Split out server to two files for testing
